setup enzyme, basic tests
